+++ app/models/consumer.rb
@@ -1,11 +1,7 @@
 class Consumer < ApplicationRecord
   has_one :user, as: :profile
 
-  validates :first_name,
-            presence: true
-
-  validates :last_name,
-            presence: true
+  validates :first_name, :last_name, presence: true
 
   def serializer
     ConsumerSerializer