+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Note: Every project begins on version 0.1.0. Clear out the log below and start anew. Remove this message once completed. -->
 
+## [0.5.0] - 2019-08-23
+### Added 
+- Password edit view
+- Separate api controller created
+- Login specs
+- Rails 5.2 config files
+
 ## [0.4.1] - 2019-07-08
 ### Fixed
 - Fixes Swagger schema validation for swagger security requirements